fix: test

+++ packages/vscode-extension/test/suite/envChecker/cases/dotnet.ts
@@ -41,7 +41,9 @@ function createTestChecker(
 async function cleanup() {
     // fs-extra.remove() does nothing if the file does not exist.
     await fs.remove(dotnetConfigPath);
-    await fs.rmdir(dotnetPrivateInstallPath);
+    if (fs.existsSync(dotnetPrivateInstallPath)) {
+      await fs.rmdir(dotnetPrivateInstallPath);
+    }
     await fs.remove(dotnetPrivateInstallPath);
 }
